+++ frontend/lost/src/actions/statistics/statistics_api.jsx
@@ -7,3 +7,9 @@ export const usePersonalStatistics = () => {
         axios.get(`${API_URL}/statistics/personal`).then((res) => res.data),
     )
 }
+
+export const useDesignerStatistics = () => {
+    return useMutation(() =>
+        axios.get(`${API_URL}/statistics/designer`).then((res) => res.data),
+    )
+}