+++ CHANGELOG.md
@@ -24,7 +24,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added sia auto save feature
 - Cronjob that removes annotations that are not assigned to any image
 - ExamplePipes:
-  - pack_dataset -> Will gather all images and a dataset file in csv and parquet format and provide it by via download 
+  - pack_dataset -> Will gather all images and a dataset file in csv and parquet format and provide it by via download
+- Dashboard:
+  - Personal statistics for annotator and designer roles
+- Statistics:
+  - Designer statistics for designers - includes all annotation activities of the users of the designer's pipelines
 ## Changed
 - Use fsspec for filesystem abstraction
   * SIA: Do not send any image urls to frontend